* ajc-out-of-place-apply:
  This commit is necessary because mypy requires explicit return of None
  ENH: enable out-of-place filtering for each filter
  ENH: replace current out-of-place apply with transform, add ability to do out-of-place apply
  Extricate Stack from filter
  Extricate Stack from spot detection.
  Extricate Stack from Segmentation.
  convert codebook retrieval to use jsonpathrw
  [easy] environment variable to keep data from a iss run. (#278)
  [easy] Move squeeze and un_squeeze to ImageStack (#277)

+++ starfish/test/full_pipelines/cli/test_iss.py
@@ -6,16 +6,28 @@ import subprocess
 import sys
 import tempfile
 import unittest
+from typing import Sequence
+
+import jsonpath_rw
 
 from starfish.util import clock
 import pytest
 
 
-def get_codebook(tempdir):
-    json_filepath = os.path.join(tempdir, "formatted", "experiment.json")
+def get_jsonpath_from_file(json_filepath_components: Sequence[str], jsonpath: str):
+    """
+    Given a series of filepath components, join them to find a file <FILE>.  Open that file, and locate a specific value
+    in the json structure <PATH>.  Join the directory path of <FILE> and <PATH> and return that.
+
+    For example, if json_path_components is ["/tmp", "formatted", "experiment.json"] and jsonpath is
+    "$['hybridization']", this method will open /tmp/formatted/experiment.json, decode that as a json document, and
+    locate the value of the key 'hybridization'.  It will return /tmp/formatted/XXX, where XXX is the value of the key.
+    """
+    json_filepath = os.path.join(*json_filepath_components)
+    dirname = os.path.dirname(json_filepath)
     with open(json_filepath, "r") as fh:
         document = json.load(fh)
-        return os.path.join(tempdir, "formatted", document['codebook'])
+        return os.path.join(dirname, jsonpath_rw.parse(jsonpath).find(document)[0].value)
 
 
 @pytest.mark.skip('long test')
@@ -46,17 +58,40 @@ class TestWithIssData(unittest.TestCase):
         ],
         [
             "starfish", "filter",
-            "--input", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "registered", "experiment.json"),
-            "--output", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered"),
+            "--input", lambda tempdir, *args, **kwargs: get_jsonpath_from_file(
+                [tempdir, "registered", "experiment.json"],
+                "$['hybridization_images']",
+            ),
+            "--output", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "hybridization.json"),
+            "white_tophat",
+            "--disk-size", "15",
+        ],
+        [
+            "starfish", "filter",
+            "--input", lambda tempdir, *args, **kwargs: get_jsonpath_from_file(
+                [tempdir, "formatted", "experiment.json"],
+                "$['auxiliary_images']['nuclei']",
+            ),
+            "--output", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "nuclei.json"),
+            "white_tophat",
+            "--disk-size", "15",
+        ],
+        [
+            "starfish", "filter",
+            "--input", lambda tempdir, *args, **kwargs: get_jsonpath_from_file(
+                [tempdir, "formatted", "experiment.json"],
+                "$['auxiliary_images']['dots']",
+            ),
+            "--output", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "dots.json"),
             "white_tophat",
             "--disk-size", "15",
         ],
         [
             "starfish", "detect_spots",
-            "--input", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "experiment.json"),
+            "--input", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "hybridization.json"),
             "--output", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "results"),
             "gaussian_spot_detector",
-            "--blobs-image-name", "dots",
+            "--blobs-stack", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "dots.json"),
             "--min-sigma", "4",
             "--max-sigma", "6",
             "--num-sigma", "20",
@@ -64,7 +99,9 @@ class TestWithIssData(unittest.TestCase):
         ],
         [
             "starfish", "segment",
-            "-i", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "experiment.json"),
+            "--hybridization-stack", lambda tempdir, *args, **kwargs: os.path.join(
+                tempdir, "filtered", "hybridization.json"),
+            "--nuclei-stack", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "filtered", "nuclei.json"),
             "-o", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "results", "regions.geojson"),
             "watershed",
             "--dapi-threshold", ".16",
@@ -82,7 +119,10 @@ class TestWithIssData(unittest.TestCase):
         [
             "starfish", "decode",
             "-i", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "results", "spots.nc"),
-            "--codebook", lambda tempdir, *args, **kwargs: get_codebook(tempdir),
+            "--codebook", lambda tempdir, *args, **kwargs: get_jsonpath_from_file(
+                [tempdir, "formatted", "experiment.json"],
+                "$['codebook']",
+            ),
             "-o", lambda tempdir, *args, **kwargs: os.path.join(tempdir, "results", "spots.nc"),
             "iss",
         ],
@@ -104,7 +144,7 @@ class TestWithIssData(unittest.TestCase):
                     element(tempdir=tempdir) if callable(element) else element
                     for element in stage
                 ]
-                if cmdline[0] == 'starfish' and coverage_enabled:
+                if cmdline[0] == "starfish" and coverage_enabled:
                     coverage_cmdline = [
                         "coverage", "run",
                         "-p",
@@ -120,10 +160,11 @@ class TestWithIssData(unittest.TestCase):
 
             counts = collections.defaultdict(lambda: 0)
             for record in results:
-                counts[record['barcode']] += 1
+                counts[record["barcode"]] += 1
             tuples = [(count, barcode) for barcode, count in counts.items()]
             tuples.sort(reverse=True)
             self.assertEqual("AAGC", tuples[0][1])
             self.assertEqual("AGGC", tuples[1][1])
         finally:
-            shutil.rmtree(tempdir)
+            if os.getenv("TEST_ISS_KEEP_DATA") is None:
+                shutil.rmtree(tempdir)
